+++ src/lesson/types/lesson.type.ts
@@ -1,5 +1,5 @@
 import { Field, ID, ObjectType } from '@nestjs/graphql';
-import { StudentType } from 'src/student/student.type';
+import { StudentType } from 'src/student/types/student.type';
 
 @ObjectType('Lesson')
 export class LessonType {
